+++ cbot_sensors/launch/cbot_sensors.launch
@@ -1,0 +1,5 @@
+<?xml version="1.0"?>
+<launch>
+    <node name="sensors_node" pkg="cbot_sensors" type="sensors_node" output="screen">
+    </node>
+</launch>
